Use the address fetched from the API instead of the client-side one
